tunable, bench: 6294847

+++ src/board.cpp
@@ -816,7 +816,7 @@ void Board::undoChangeColor() {
 int Board::getEvaluation() {   
     //std::cout << "position fen " << getFenString() << std::endl;
     //std::cout << "evaluate" << std::endl;
-    int eval = int(double(nnueState.evaluate(colorToMove, __builtin_popcountll(getOccupiedBitboard()))));
+    int eval = nnueState.evaluate(colorToMove, __builtin_popcountll(getOccupiedBitboard()));
     //nnueState.fullRefresh(stateHistory.back(), stateHistory.back().kingSquares[0], stateHistory.back().kingSquares[1]);
     //assert(eval == nnueState.evaluate(colorToMove, __builtin_popcountll(getOccupiedBitboard())));
     int phase =  3 * __builtin_popcountll(stateHistory.back().pieceBitboards[Knight])
